+++ src/components/SynonymCard.js
