+++ src/components/WordCard.js
