+++ src/index.js
@@ -2,7 +2,7 @@ import React from "react";
 import ReactDOM from "react-dom";
 import "antd/dist/antd.css";
 import "./index.css";
-import Home from "./Home";
+import Home from "./pages/Home";
 
 ReactDOM.render(
   <React.StrictMode>
